requirements.txt
Django
django-address==0.2.8
django-cors-headers==4.3.0
django-crispy-forms==1.14.0
django-extensions==3.2.3
django-filter==22.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
mysqlclient
Markdown==3.4.1
pymysql     